+++ ewm/src/main/java/ru/practicum/compilations/service/publ/CompilationPublicService.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface CompilationPublicService {
     List<CompilationDto> getCompilations(Boolean pinned, Integer from, Integer size);
 
-    CompilationDto getCompilationById(@PathVariable Integer compId);
+    CompilationDto getCompilationById(@PathVariable Long compId);
 }
