+++ ewm/src/main/java/ru/practicum/events/dto/publ/EventFullDto.java
@@ -27,7 +27,7 @@ public class EventFullDto {
     @Schema(description = "Категория", example = "{\"id\": 1, \"name\": \"Концерты\"}")
     private CategoryDto category;
     @Schema(description = "Количество одобренных заявок на участие в данном событии")
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @Schema(description = "Дата и время создания события (в формате yyyy-MM-dd HH:mm:ss)")
     private LocalDateTime createdOn;
@@ -37,7 +37,7 @@ public class EventFullDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @Schema(description = "Дата и время на которые намечено событие (в формате yyyy-MM-dd HH:mm:ss)")
     private LocalDateTime eventDate;
-    private Integer id;
+    private Long id;
     @NotNull
     @Schema(description = "Пользователь (краткая информация)", example = "{\"id\": 3, \"name\": \"Фёдоров Матвей\"}")
     private UserShortDto initiator;
@@ -48,7 +48,7 @@ public class EventFullDto {
     @Schema(description = "Нужно ли оплачивать участие")
     private Boolean paid;
     @Schema(description = "Ограничение на количество участников. Значение 0 - означает отсутствие ограничения")
-    private Integer participantLimit = 0;
+    private Long participantLimit = 0L;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @Schema(description = "Дата и время публикации события (в формате yyyy-MM-dd HH:mm:ss)")
     private LocalDateTime publishedOn;
@@ -60,5 +60,5 @@ public class EventFullDto {
     @Schema(description = "Заголовок")
     private String title;
     @Schema(description = "Количество просмотрев события")
-    private Integer views;
+    private Long views;
 }
